src/cfs/low_volume_skus_feed_generation/requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
db-dtypes==1.0.5
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-bigquery==3.6.0
google-cloud-bigquery-datatransfer==3.11.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
httplib2==0.21.0
idna==3.4
numpy==1.24.2
packaging==23.0
pandas==1.5.3
proto-plus==1.22.2
protobuf==4.22.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.14
